Text Spacing Fix
Outlook 2013 was a) not using margins, b) not building margin when a paragraph ended in a link.
Solution: Class to set margins to 0px in other clients and then used foundation inky spacer tags.
Now distances between paragraphs are like in the template in all clients. (Except for Lotus notes 8.5 which is annoying anyway.)